requirements.txt
Flask
torch
diffusers
Pillow
transformers